plan, executor: improve hash join to support all join types
