+++ plan/task.go
@@ -168,7 +168,7 @@ func (p *PhysicalIndexJoin) getCost(lCnt float64) float64 {
 
 func (p *PhysicalHashJoin) getCost(lCnt, rCnt float64) float64 {
 	smallTableCnt := lCnt
-	if p.SmallTable == 1 {
+	if p.SmallChildIdx == 1 {
 		smallTableCnt = rCnt
 	}
 	if smallTableCnt <= 1 {
